requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-filter==2.4.0
django-heroku==0.3.1
django-pure-pagination==0.3.0
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==3.3.0
isort==5.6.4
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.19.3
pandas==1.1.5
plotly==4.14.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
